+++ test_mod.rs
@@ -1,0 +1,4 @@
+use silk_parser::Parser;
+let source = "result = 10 % 3";
+let program = Parser::parse(source).unwrap();
+println!("Parsed OK: {:?}", program);